Fixed NEL calls to updated dl4j API

+++ texoo-entity-linking/src/main/java/de/datexis/index/encoder/EntityEncoder.java
@@ -49,7 +49,7 @@ public class EntityEncoder extends Encoder {
   }
   
   @Override
-  public int getVectorSize() {
+  public long getVectorSize() {
     if(strategy.equals(Strategy.NAME)) return parvec.getLayerSize();
     else if(strategy.equals(Strategy.NAME_CONTEXT)) return parvec.getLayerSize() * 2;
     else throw new IllegalArgumentException("invalid strategy");
